Add `bun info` command as alias for `bun pm view`
